# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestionar múltiples sesiones de Telegram via MTProto",
        "title": "Telegram API",
        "contact": {},
        "version": "0.3.0"
    },
    "host": "localhost:7789",
    "basePath": "/api/v1",
    "paths": {
...
        "/messages/{jobId}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Consulta estado de un mensaje enviado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Estado del mensaje",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MessageJob"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.MessageJob": {
            "description": "Estado detallado del mensaje",
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "media_url": {
                    "type": "string"
                },
                "send_at": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.MessageStatus"
                        }
                    ],
                    "example": "sent"
                },
                "text": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "example": "@username"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.MessageType"
                        }
                    ],
                    "example": "text"
                }
            }
        },
...
        "domain.MessageStatus": {
            "type": "string",
            "enum": [
                "pending",
                "scheduled",
                "sending",
                "sent",
                "failed"
            ],
            "x-enum-varnames": [
                "MessageStatusPending",
                "MessageStatusScheduled",
                "MessageStatusSending",
                "MessageStatusSent",
                "MessageStatusFailed"
            ]
        },
        "domain.MessageType": {
            "type": "string",
            "enum": [
                "text",
                "photo",
                "video",
                "audio",
                "file"
            ],
            "x-enum-varnames": [
                "MessageTypeText",
                "MessageTypePhoto",
                "MessageTypeVideo",
                "MessageTypeAudio",
                "MessageTypeFile"
            ]
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.MetaResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/handler.ErrorResponse"
                },
                "meta": {
                    "$ref": "#/definitions/handler.MetaResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}